feat: update image registry
