+++ libs/domains/organizations/feature/src/index.ts
@@ -9,6 +9,7 @@ export * from './lib/git-token-create-edit-modal/git-token-create-edit-modal'
 export * from './lib/helm-repository-create-edit-modal/helm-repository-create-edit-modal'
 export * from './lib/git-token-list/git-token-list'
 export * from './lib/container-registry-create-edit-modal/container-registry-create-edit-modal'
+export * from './lib/container-registry-form/container-registry-form'
 export * from './lib/hooks/use-add-credit-card/use-add-credit-card'
 export * from './lib/hooks/use-add-credit-code/use-add-credit-code'
 export * from './lib/hooks/use-annotations-groups/use-annotations-groups'
